Fix broken lines caused by missing series values
In the case of multi series and multi builds, if there
are missed values on some builds with a couple of series
broken lines would be plotted.
Here generatePlot can the second round scan and fix missed
values with valid values on their former builds if there
would be any.